Fix warning S2184 in Polly.Spec

+++ test/Polly.Specs/Retry/WaitAndRetrySpecs.cs
@@ -1098,7 +1098,7 @@ public class WaitAndRetrySpecs : IDisposable
         attemptsInvoked.Should().Be(1);
 
         watch.Elapsed.Should().BeLessThan(retryDelay);
-        watch.Elapsed.Should().BeCloseTo(shimTimeSpan, precision: TimeSpan.FromMilliseconds((int)shimTimeSpan.TotalMilliseconds / 2));  // Consider increasing shimTimeSpan, or loosening precision, if test fails transiently in different environments.
+        watch.Elapsed.Should().BeCloseTo(shimTimeSpan, precision: TimeSpan.FromMilliseconds(shimTimeSpan.TotalMilliseconds / 2));  // Consider increasing shimTimeSpan, or loosening precision, if test fails transiently in different environments.
     }
 
     [Fact]
